requirements.txt
# llm and langchain
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
openai>=1.0.0

# embeddings
sentence-transformers>=2.2.0

# vector store
chromadb>=0.4.0

# pdf parsing
PyMuPDF>=1.23.0

# backend
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0

# frontend
streamlit>=1.30.0

# utilities
numpy>=1.24.0
requests>=2.28.0